Do not override subclaim type
