+++ modules/dependents_benefits/app/models/dependents_benefits/parent_dependency.rb
@@ -1,9 +1,13 @@
 # frozen_string_literal: true
 
+require 'dependents_benefits/claim_behavior'
+
 module DependentsBenefits
   # DependentsBenefit 21-509 Active::Record
   # @see app/model/saved_claim
-  class ParentDependency < DependentsBenefits::SavedClaim
+  class ParentDependency < ::SavedClaim
+    include DependentsBenefits::ClaimBehavior
+
     # DependentsBenefit Form ID
     FORM = DependentsBenefits::PARENT_DEPENDENCY
