+++ modules/dependents_benefits/spec/controllers/dependents_benefits/v0/claims_controller_spec.rb
@@ -59,7 +59,15 @@ RSpec.describe DependentsBenefits::V0::ClaimsController do
       it 'creates saved claims' do
         expect do
           post(:create, params: test_form, as: :json)
-        end.to change(DependentsBenefits::SavedClaim, :count).by(3)
+        end.to change(
+          DependentsBenefits::SavedClaim, :count
+        ).by(1)
+          .and change(
+            DependentsBenefits::AddRemoveDependent, :count
+          ).by(1)
+          .and change(
+            DependentsBenefits::SchoolAttendanceApproval, :count
+          ).by(1)
       end
 
       it 'calls ClaimProcessor with correct parameters' do
