+++ modules/dependents_benefits/spec/lib/dependents_benefits/generators/dependent_claim_generator_spec.rb
@@ -9,9 +9,9 @@ RSpec.describe DependentsBenefits::Generators::DependentClaimGenerator, type: :m
   let(:generator) { described_class.new(form_data, parent_id) }
 
   before do
-    allow(generator).to receive(:claim_class).and_return(DependentsBenefits::SavedClaim)
-
     allow_any_instance_of(SavedClaim).to receive(:pdf_overflow_tracking)
+
+    allow(generator).to receive(:claim_class).and_return(DependentsBenefits::SavedClaim)
   end
 
   describe 'initialization' do
